Simplify adding new interpretation
