+++ nomic-bot/src/universe/interpreter/createComment.js
@@ -1,0 +1,7 @@
+import { sendCommentMessage } from '../../utils/github'
+import { addInterpretation, createComment } from './index';
+
+
+addInterpretation(createComment.type, ({ comment }, { commentsUrl }) => {
+    return sendCommentMessage(commentsUrl, comment);
+ })
