+++ nomic-bot/src/universe/interpreter/index.js
@@ -1,0 +1,9 @@
+// Import your interpreter here.
+import './updatePlayer';
+import './createComment';
+
+
+// Exports
+export * from '../actions';
+export { addInterpretation } from './utils';
+export default from './utils';
